Add exports, separate extensions
