+++ mod.ts
@@ -4,19 +4,13 @@ export type {
   Thresholds,
 } from "./types.ts";
 
-export {
-  prettyBenchmarkResult,
-} from "./pretty_benchmark_result.ts";
-export type {
-  prettyBenchmarkResultOptions,
-} from "./pretty_benchmark_result.ts";
+export { calculateExtraMetrics, calculateStdDeviation } from "./common.ts";
 
-export {
-  prettyBenchmarkProgress,
-} from "./pretty_benchmark_progress.ts";
-export type {
-  prettyBenchmarkProgressOptions,
-} from "./pretty_benchmark_progress.ts";
+export { prettyBenchmarkResult } from "./pretty_benchmark_result.ts";
+export type { prettyBenchmarkResultOptions } from "./pretty_benchmark_result.ts";
+
+export { prettyBenchmarkProgress } from "./pretty_benchmark_progress.ts";
+export type { prettyBenchmarkProgressOptions } from "./pretty_benchmark_progress.ts";
 
 export {
   prettyBenchmarkDown,
@@ -31,3 +25,8 @@ export type {
   ColumnDefinition,
   GroupDefinition,
 } from "./pretty_benchmark_down.ts";
+
+export {prettyBenchmarkHistory, calculateThresholds} from './pretty_benchmark_history.ts';
+export type {BenchmarkHistory, BenchmarkHistoryItem, BenchmarkHistoryRunItem, Delta, prettyBenchmarkHistoryOptions} from './pretty_benchmark_history.ts';
+
+export {deltaColumn, deltaProgressRowExtra, deltaResultInfoCell, historyColumns} from './history_extensions.ts';
