+++ pretty_benchmark_history.ts
@@ -1,12 +1,5 @@
-import { BenchmarkResult, BenchmarkRunResult, bench, runBenchmarks, colors } from "./deps.ts";
-import { calculateExtraMetrics, calculateStdDeviation, stripColor } from "./common.ts";
-import { BenchIndicator, Thresholds, Threshold } from "./types.ts";
-import { rtime } from "./utils.ts";
-// TODO move these to separate file
-// TODO export calcStd, calcExtraMetrics
-import { prettyBenchmarkProgress, prettyBenchmarkProgressOptions } from "./pretty_benchmark_progress.ts";
-import { prettyBenchmarkResult, prettyBenchmarkResultOptions } from "./pretty_benchmark_result.ts";
-import { prettyBenchmarkDown, ColumnDefinition } from "./pretty_benchmark_down.ts";
+import type { BenchmarkResult, BenchmarkRunResult } from "./deps.ts";
+import type { Thresholds, Threshold } from "./types.ts";
 
 export interface prettyBenchmarkHistoryOptions<T = unknown, K = unknown> {
     runExtras?: (runResult: BenchmarkRunResult) => K;
@@ -46,7 +39,7 @@ export interface Delta {
     amount: number;
 }
 
-type DeltaKey<T = unknown> = (keyof T | "measuredRunsAvgMs" | "totalMs");
+export type DeltaKey<T = unknown> = (keyof T | "measuredRunsAvgMs" | "totalMs");
 
 export class prettyBenchmarkHistory<T = unknown, K=unknown> {
     private data!: BenchmarkHistory<T, K>;
@@ -258,191 +251,3 @@ export function calculateThresholds<T, K>(history: prettyBenchmarkHistory<T, K>,
 
     return thresholds;
 }
-
-export function deltaProgressRowExtra(history: prettyBenchmarkHistory) {
-    return (result: BenchmarkResult, options?: prettyBenchmarkProgressOptions) => {
-        let deltaString = getCliDeltaString(history, result);
-
-        if(options?.nocolor) {
-            deltaString = stripColor(deltaString);
-        }
-
-        return ` [${deltaString}]`;
-    };
-}
-
-export function deltaResultInfoCell(history: prettyBenchmarkHistory) {
-    return (result: BenchmarkResult, options?: prettyBenchmarkResultOptions) => {
-        let deltaString = getCliDeltaString(history, result);
-
-        if(options?.nocolor) {
-            deltaString = stripColor(deltaString);
-        }
-
-        return `  ${deltaString}`;
-    };
-}
-
-function getCliDeltaString(history: prettyBenchmarkHistory, result: BenchmarkResult) {
-    const delta = history.getDeltaForBenchmark(result);
-    let deltaString = `${colors.gray(" ▪   no history  ▪ ".padEnd(19))}`;
-    if(delta) {
-        const perc = (delta.measuredRunsAvgMs.percent * 100).toFixed(0);
-        const diff = rtime(Math.abs(delta.measuredRunsAvgMs.amount));
-        
-        if(delta.measuredRunsAvgMs.amount > 0) {
-            deltaString = `${colors.red(` ▲ ${`+${perc}`.padStart(4)}% (${diff.padStart(6)}ms)`)}`;
-        } else {
-            deltaString = `${colors.green(` ▼ ${perc.padStart(4)}% (${diff.padStart(6)}ms)`)}`;
-        }
-    }
-
-    return deltaString;
-}
-
-export function deltaColumn<T = unknown>(history: prettyBenchmarkHistory<T>, options?: {key: DeltaKey<T>}): ColumnDefinition {
-    const workingKey = options?.key || "measuredRunsAvgMs";
-
-    return {title: `Change in ${options?.key || "average"}`, formatter: (result, cd) => {
-        const delta = history.getDeltaForBenchmark(result, [workingKey]);
-        if(delta) {
-            const perc = (delta[workingKey as string].percent * 100).toFixed(0);
-            const diff = rtime(Math.abs(delta.measuredRunsAvgMs.amount));
-            
-            const notSpaceChar = " ";
-            const smallSpace = " "; 
-            if(delta[workingKey as string].amount > 0) {
-                return `🔺 ${`+${perc}`.padStart(4,notSpaceChar)}% (${diff.padStart(6)}ms)`;
-            } else {
-                return `🟢${smallSpace} ${perc.padStart(4, notSpaceChar)}% (${diff.padStart(6)}ms)`;
-            }
-        }
-        return "-";
-    }};
-}
-
-export function historyColumns<T = unknown>(history: prettyBenchmarkHistory<T>, options?: {key?: DeltaKey<T>, titleFormatter?: (date: Date, id?: string) => string}): ColumnDefinition[] {
-    if(history.getData().history.length === 0){
-        return [];
-    }
-
-    const dateFormatter = (d: Date) => {
-        return d.toISOString().split("T").join('<br/>').replace(/Z/, "");
-    };
-
-    return history.getData().history.map(run => {        
-        const parsedDate = new Date(run.date);
-        return {
-            title: (typeof options?.titleFormatter === "function" ? options.titleFormatter(parsedDate, run.id) : run.id ?? dateFormatter(parsedDate)),
-            toFixed: 4,
-            formatter: (result: BenchmarkResult) => {
-                if(!run.benchmarks[result.name]) {
-                    return "-";
-                }
-
-                const workingKey = options?.key ?? "measuredRunsAvgMs";
-
-                if(workingKey === "measuredRunsAvgMs" || workingKey === "totalMs") {
-                    return (run.benchmarks[result.name] as any)[workingKey as any] || "-";
-                } else {
-                    return run.benchmarks[result.name].extras?.[workingKey] || "-";
-                }
-            }
-        };
-    });
-}
-
-function example() {
-
-    bench({
-        name: "historic",
-        func(b) {
-            b.start();
-            for (let i = 0; i < 1e3; i++) {
-                const NPeP = Math.random() === Math.random();
-            }
-            b.stop();
-        },
-        runs: 500
-    });
-
-    bench({
-        name: "x3#14",
-        func(b) {
-            b.start();
-            for (let i = 0; i < 1e5; i++) {
-                const NPeP = Math.random() === Math.random();
-            }
-            b.stop();
-        },
-        runs: 1000
-    });
-    bench({
-        name: "MZ/X",
-        func(b) {
-            b.start();
-            for (let i = 0; i < 1e5; i++) {
-                const NPeP = Math.random() === Math.random();
-            }
-            b.stop();
-        },
-        runs: 1000
-    });
-
-    bench({
-        name: "MZ/T",
-        func(b) {
-            b.start();
-            for (let i = 0; i < 1e5; i++) {
-                const NPeP = Math.random() === Math.random();
-            }
-            b.stop();
-        },
-        runs: 1000
-    });
-    
-    let prevString;
-    try {
-        prevString = JSON.parse(Deno.readTextFileSync('./benchmarks/historicx.json'));
-    } catch {
-        console.warn('⚠ cant read file');
-    }
-
-    const history = new prettyBenchmarkHistory({ 
-        saveIndividualRuns: false,
-        minRequiredRuns: 100,
-        easeOnlyHrTime: false,
-        strict: {noAddition: false, noRunsCountChange: false, noRemoval: false},
-        benchExtras: (r: BenchmarkResult) => ({r: r.name, ...calculateExtraMetrics(r), std: calculateStdDeviation(r)}),
-        runExtras: (rr: BenchmarkRunResult) => ({dv: Deno.version, f: rr.filtered})
-    }, prevString);
-
-    // console.log(JSON.stringify(historic.getData()));
-
-    const inds: BenchIndicator[] = [
-        {benches: /historic/, modFn: _ => "👃"}
-    ];
-    const thds = calculateThresholds(history);
-
-    runBenchmarks({silent: true}, prettyBenchmarkProgress({rowExtras: deltaProgressRowExtra(history),indicators: inds, nocolor: false, thresholds: thds}))
-        // TODO defaultColumns to func, dont get avg, total, just name, maybe runs
-        .then(prettyBenchmarkDown(md => {Deno.writeTextFileSync("./benchmarks/hmdx.md", md)}, {columns: [{title: 'Name', propertyKey: 'name'}, ...historyColumns(history), {title: 'Average (ms)', propertyKey: 'measuredRunsAvgMs', toFixed: 4}, deltaColumn(history)]})) // historicColumn
-        .then(prettyBenchmarkResult({infoCell: deltaResultInfoCell(history), nocolor: false, thresholds: thds, parts: {threshold: true, graph: true}}))
-        .then((results: BenchmarkRunResult) => {
-
-
-            // console.log(historic.getDeltasFrom(results, "max"))
-            history.addResults(results);
-            // console.log(history.getDataString());
-            
-            // console.log(historic.getDeltasFrom(results));
-            
-            // Deno.writeTextFileSync("./benchmarks/historicx.json", historic.getDataString());
-        });
-
-    return;
-}
-
-// deno run --allow-read --allow-write --allow-hrtime .\pretty_benchmark_historic.ts
-
-example();
